global commit, going to get better at making them smaller

+++ kernel/interupt_handling.h
@@ -4,7 +4,7 @@
 #define IDT_TYPE_INTERRUPT 0xE
 #define IDT_TYPE_TRAP 0xF
 
-void memset(void* target, uint64_t c, uint64_t n);
+void memset(void* target, int c, int n);
 
 /**
  * Set an interrupt handler for the given interrupt number.
